requirements.txt
pandas
altair
pytorch
'python3 -m spacy download de_core_news_sm'